add `blitz generate mutation` (minor)
