+++ packages/generator/templates/mutations/delete__ModelName__.ts
